Refactor: rename section dir and add card index
