+++ src/components/Card/index.tsx
@@ -1,0 +1,1 @@
+export { IdeaCard } from "./IdeaCard";
