+++ src/components/Section/TodayIdea.tsx
